+++ lua/config/autocmds.lua
@@ -20,7 +20,7 @@ end
 vim.api.nvim_create_autocmd({ "VimEnter" }, { callback = open_nvim_tree })
 
 -- auto change CRLF to LF
-vim.api.nvim_create_autocmd("BufReadPost", {
+vim.api.nvim_create_autocmd("BufWrite", {
   pattern = "*",
   callback = function()
     if vim.bo.fileformat == "dos" then